[30626] Add Open Doors Method
This method sets the open doors bool to true in the animator
