+++ Assets/Scripts/Level/Sliding_Doors/SlidingDoors.cs.meta
